refactor: update explore and search screens with new adapters
