+++ app/src/main/java/com/karrar/movieapp/ui/search/adapters/MediaSearchCardAdapter.kt
@@ -9,7 +9,7 @@ import com.karrar.movieapp.ui.search.mediaSearchUIState.MediaUIState
 
 class MediaSearchCardAdapter(listener: SearchViewModel)
     : BasePagingAdapter<MediaUIState>(MediaSearchCardComparator, listener){
-    override val layoutID: Int = R.layout.item_media_card_search
+    override val layoutID: Int = R.layout.list_media_item
 
     object MediaSearchCardComparator : DiffUtil.ItemCallback<MediaUIState>(){
         override fun areItemsTheSame(oldItem: MediaUIState, newItem: MediaUIState) =
